+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: Lmatkows <lmatkows@student.42perpignan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 18:06:27 by Lmatkows          #+#    #+#             */
-/*   Updated: 2024/10/24 18:45:27 by Lmatkows         ###   ########.fr       */
+/*   Updated: 2024/10/31 10:31:04 by Lmatkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void	*adr, int c, int len)
+#include "libft.h"
+
+void	*ft_memset(void	*adr, int c, size_t len)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	i = 0;
